Added splitscreen support
Unfortunately, this had to come at the cost of the pixel perfect camera, which absolutely does NOT play nice with splitscreen, as it demands absolute control over the camera's viewport rect. As such, we will be forced to use the base camera, and unless we make a dedicated singleplayer mode, we won't be bringing the pixel-perfect camera back anytime soon.